M4 f2 savings accounts

+++ project/sql/006_create_table_transactions.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS `Transactions` (
     `id`                    INT NOT NULL AUTO_INCREMENT,
     `account_source`        INT NOT NULL,
-    `account_destination`   INT NOT NULL,
-    `balance_change`        INT NOT NULL,
+    `account_destination`   DECIMAL(12, 2) NOT NULL,
+    `balance_change`        DECIMAL(12, 2) NOT NULL,
     `transaction_type`      VARCHAR(12),
     `memo`                  VARCHAR(150) DEFAULT '',
     `transaction_time`      TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
